feat(steps): adding new step "log files"
