+++ src/constants/constants.js
@@ -1,0 +1,27 @@
+export const platforms = [
+  'Ansible',
+  'AzureResourceManager',
+  'Buildah',
+  'CloudFormation',
+  'DockerCompose',
+  'Dockerfile',
+  'GRPC',
+  'GoogleDeploymentManager',
+  'Kubernetes',
+  'OpenAPI',
+  'Terraform'
+]
+
+export const outputFormats = [
+  'all',
+  'asff',
+  'csv',
+  'cyclonedx',
+  'glsast',
+  'html',
+  'json',
+  'junit',
+  'pdf',
+  'sarif',
+  'sonarqube'
+]
