+++ src/pages/Main/MainSteps/MainStep3.jsx
@@ -2,100 +2,35 @@ import Box from '../../../components/Box/Box'
 import RadioGroup from '../../../components/RadioGroup/RadioGroup'
 import Input from '../../../components/Input/Input'
 
-const MainStep3 = () => {
+const MainStep4 = () => {
   return (
     <>
-      <Box title="Minimal UI">
+      <Box title="Log Level">
         <RadioGroup
-          flag="minimal-ui"
-          defaultIndex={1}
+          flag="log-level"
+          defaultIndex={2}
           items={[
-            {
-              label: 'Enable',
-              flag: 'minimal-ui',
-              flagAux: 'enable'
-            },
-            { label: 'Disable', flag: 'minimal-ui', flagAux: 'disable' }
+            { label: 'TRACE', value: 'TRACE', flagAux: 'TRACE', flag: 'log-level' },
+            { label: 'DEBUG', value: 'DEBUG', flagAux: 'DEBUG', flag: 'log-level' },
+            { label: 'INFO (default)', value: 'INFO', flagAux: 'INFO', flag: 'log-level' },
+            { label: 'WARN', value: 'WARN', flagAux: 'WARN', flag: 'log-level' },
+            { label: 'ERROR', value: 'ERROR', flagAux: 'ERROR', flag: 'log-level' },
+            { label: 'FATAL', value: 'FATAL', flagAux: 'FATAL', flag: 'log-level' }
           ]}
         />
       </Box>
-      <Box title="No Progress">
-        <RadioGroup
-          flag="no-progress"
-          defaultIndex={1}
-          items={[
-            {
-              label: 'Enable',
-              flag: 'no-progress',
-              flagAux: 'enable'
-            },
-            { label: 'Disable', flag: 'no-progress', flagAux: 'disable' }
-          ]}
-        />
-      </Box>
-      <Box title="Preview Lines">
-        <RadioGroup
-          flag="preview-lines"
-          defaultIndex={1}
-          items={[
-            {
-              label: 'Enable',
-              flag: 'preview-lines',
-              flagAux: null,
-              children: (
-                <Input
-                  flag="preview-lines"
-                  flagAux={null}
-                  placeholder={'Number of lines to be display in CLI results (min: 1, max: 30)'}
-                  label={'Nº of Lines'}
-                  description={'Number of lines to be display in CLI results (min: 1, max: 30)'}
-                  defaultValue={'3'}
-                />
-              )
-            },
-            { label: 'Disable', flag: 'preview-lines', flagAux: 'disable' }
-          ]}
-        />
-      </Box>
-      <Box title="No Color">
-        <RadioGroup
-          flag="no-color"
-          defaultIndex={1}
-          items={[
-            {
-              label: 'Enable',
-              flag: 'no-color',
-              flagAux: 'enable'
-            },
-            { label: 'Disable', flag: 'no-color', flagAux: 'disable' }
-          ]}
-        />
-      </Box>
-      <Box title="Output Style">
-        <RadioGroup
-          flag="custom-cli-01"
-          defaultIndex={0}
-          items={[
-            {
-              flag: 'custom-cli-01',
-              flagAux: 'default',
-              label: 'Default'
-            },
-            {
-              flag: 'custom-cli-01',
-              flagAux: 'silent',
-              label: 'Silent'
-            },
-            {
-              flag: 'custom-cli-01',
-              flagAux: 'verbose',
-              label: 'Verbose'
-            }
-          ]}
+
+      <Box title="Log Path (output file)">
+        <Input
+          flag={'log-path'}
+          flagAux={'local'}
+          label="Log Output Path"
+          placeholder="Complete path (including filename) to store the output log"
+          description={'Example: c:\\scans-output\\scan.log'}
         />
       </Box>
     </>
   )
 }
 
-export default MainStep3
+export default MainStep4
